+++ src/components/footer.jsx
@@ -1,0 +1,7 @@
+export default function Footer(){
+  return (
+    <footer className="bg-slate-900 text-slate-200 text-center px-4 py-4 border-t border-slate-800">
+      <p className="opacity-75">© 2026 - CloudBook Derechos reservados</p>
+    </footer>
+  )
+}