+++ src/components/main-container.jsx
@@ -1,0 +1,7 @@
+export default function MainContainer(props){
+  return(
+    <main className="min-h-dvh grid grid-cols">
+      {props.children}
+    </main>
+  )
+}